requirements.txt
absl-py
langdetect
nltk
immutabledict
spacy
emoji
syllapy
unicodedata2
openai
pyyaml
pandas
pyarrow
tqdm
